Update patch set 3
Patch Set 3: Code-Review+1

Code change LGTM, but I think the commit message does not explain why we're making this change. I would say that the commit message should say something like: "Currently, getnameinfo() will never issue a reverse DNS lookup if the first byte of the IPv6 address is 0x00. This means it is not possible to do a reverse DNS lookup for a NAT64 address if the NAT64 prefix is the well-known prefix. Make this possible by treating the well-known prefix specially. This is not needed for unicast NAT64 prefixes because they do not start with 0x00."

Patch-set: 3
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1