Update patch set 2
Patch Set 2:

The name "Android" here is the "Name" field of an ELF note. It's a vendor name that creates a namespace for the type of the note, which is vendor-specific. ("Android", 1) and ("LinuxBionic", 1) are really the same kind of note, so I think it makes sense to use "Android" for both of them.

There isn't a Linux" vendor AFAIK, but there are vendors like "GNU", "NetBSD", etc.

FWIW: there is a script ndk/parse_elfnote.py that parses this Android version note:

https://android.googlesource.com/platform/ndk/+/master/parse_elfnote.py

Patch-set: 2
CC: Gerrit User 1229339 <1229339@85c56323-6fa9-3386-8a01-6480fb634889>