Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > why can't you sscanf? i thought that only allocated for floats?
> 
> Sorry, this is a WIP, I just forgot to mark it as such.
> 
> Do you mean to only have a single buffer for the lo and high pointer and then using sscanf?

oh, from the diff in email it looked like this was still for /proc/self/stat, not maps. yeah, dealing with maps is why i couldn't see a good way to do this. (not for the first time, i find myself wishing i had a system call to ask about a given VMA...)

Patch-set: 2