Introduce android_set_allocation_limit() function.
This function allows programs to set an upper bound on the total
size of all malloc() allocations. This is useful for programs
such as audioextractor and mediaserver which need to set such a
limit as a security mitigation. Currently these programs are using
setrlimit(RLIMIT_AS) which isn't exactly what these programs want
to control. RLIMIT_AS is also problematic under sanitizers which
allocate large amounts of address space as shadow memory, and is
especially problematic under shadow call stack, which requires 16MB
of address space per thread.

Bug: 118642754
Change-Id: I735403c4d2c87f00fb2cdef81d00af0af446b2bb