Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The static libc still needs to work on pre-Q.

Of course. That goes back to our discussion on bug 119390260.

Consider this for illustration purposes only - I'm finding all the places I need to touch. There is no future I'm aware of where we don't read the tzdata file out of /apex for the dynamic usecase.

Is there a way to identify in code (or preprocessor) that it's the static usecase? i.e. so we can avoid trying the fallbacks and just go for the known path?

We'll probably need to keep a static copy of tzdata lying around in /system (or some new path) that we don't update. We should work out if that's better or worse than returning "UTC" behavior for everything (I haven't forgotten what you've said about trying to keep the static version working the same but I'm still processing that - it may take a while). Of course, we also can't read out of a /apex path from things like initd either.

Patch-set: 1
Reviewer: Gerrit User 1038039 <1038039@85c56323-6fa9-3386-8a01-6480fb634889>