Update patch set 1
Patch Set 1:

I don't think we'd be consuming more pages, at least not most of the times. I'm assuming you are referring to objects being pushed out to the next page by the added information stored at the beginning of a page, but given that there is already a page_info struct there, we are already eating that cost (especially for large objects). What's happening with this CL is that for small objects, we are storing 1~2 less objects per page, but that's a tiny cost given that they are small objects. Not sure if I misunderstood what you were saying. My apology if I did.

And thanks for the suggestions. I'll update the CL with how I measured memory usage.

Patch-set: 1
Reviewer: Gerrit User 1349653 <1349653@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0