Update patch set 11
Patch Set 11:

I'm not really fond of adding another function callback. Could this be done through a mallopt call? That's kind of where we are trying to put this kind of stuff so that we don't add extra functions all of the time.

Patch-set: 11
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>