Add stubs variants for bionic libs
libc, libm, and libdl are added with stubs variant. When the runtime
APEX includes the bionic libs, then dependencies to the bionic libs from
outside of the runtime APEX are satisfied by the stubs variants which
are created from the map file.

In doing so, several symbols that are referenced from other system
components outside of the runtime APEX are added with "# vndk" tag. The
symbols are available only across APEXes, but are not available to apps
(via NDK stubs) or vendors (via LLNDK stubs).

Bug: 120266448
Test: m
Change-Id: I7b8d75830c81d7d7d54e2fca21a85b3417531b47