Update patch set 1
Patch Set 1:

In retrospect, this may not be the best test... This doesn't actually assert anything on our partner devices, as these constants will be embedded within the CTS binary itself. If we really want to check the java constants, we need to write a java test, which loads the constant from JNI....

This test might still be useful, but I'll defer to enh if he wants it or not.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>