Update patch set 1
Patch Set 1: Code-Review+2

I think this is likely to make it a bit less flaky, but I can imagine that it still might be flaky. The problem is that since we run multiple tests at the same time, if we spin up or spin down a bunch of processes it might go out of bounds.

If it is still flaky, maybe you could run the test again if it fails, and fail if it fails twice in a row.

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2