${RUNTIME} is expanded to /system/bootstrap
There will be two sets of bionic libs: the ones in the runtime APEX and
the ones in /system/bootstrap/lib[64]. The former is used for all
processes launched after the runtime APEX is activated and the latter is
used for a small number of processes (such as init, vold, and apexd
itself) that are started before the APEX is ready.

Since the dynamic linker has to use different bionic libs depending on
the context, the location of the bionic libs is abstracted behind the
symbol 'RUNTIME', which will be expanded to either
/system/bootstrap or /apex/com.androd.runtime.

Right now, we are not using the bionic libs in the runtime APEX. So the
expansion is fixed to /system/bootstrap, but it will be expanded
differently for the bootstrap linker (which is /system/bin/linker) and
the ordinary linker (which is /apex/com.android.runtime/bin/linker).

Bug: http://b/120266448
Bug: http://b/119274217
Test: device boots
Test: `lsof | grep libc` shows that libc.so is from
/system/bootstrap/lib[64]

Change-Id: Ib638e9d597890e71653e87b32e5fd678c4656e8c