Add /system/bootstrap to the default search paths
For processes that are started before the runtime APEX is activated, the
bootstrap bionic libs (under /system/bootstrap/lib[64]) are used.
Therefore the path should be added to the search paths. For most of the
early processes, this is done by adding the path to
/system/etc/ld.config.<ver>.txt file.

However, the file can't be used for init; the <ver> is from the system
property ro.vndk.version, but system properties are not available until
when init runs. As a result, the dynamic linker fails to find the config
file and falls back to the hard-coded paths.

Thefore, the bootstrap path is also added to the hard-coded default
paths.

Bug: 120266448
Test: device boots
Change-Id: Icfa13ccde595c20c747a54ff020b9595a9dfe15c