Add pthread_alloc_static_tls_word_np
This function allocates a word of static TLS memory from the pthread keys
but does not actually return a key. Instead, it returns a TP-relative
offset and (for sanity-checking) the current value of the TP.

The word is guaranteed to be zero-initialized, but because we don't want
to use a thread list to clear the memory, we can only use a key that
hasn't been used yet. These words can't generally be recycled, so there's
no API for freeing them.

The only intended use currently is Golang, which needs a word of static
TLS memory to hold its "g" register.

See https://groups.google.com/forum/#!topic/golang-dev/yVrkFnYrYPE.

Bug: http://b/78026329
Bug: http://b/118381796
Test: bionic unit tests
Change-Id: I2de3ed54fe9dd682d3ea62c39e1b656caf1374b5