Protect x18 from being clobbered by vDSO on certain kernel versions.
Treble requires us to be able to run if the kernel is from an older
release than the system partition. This means that it's possible for
us to be running on a device with a kernel whose vDSO clobbers the
x18 register (which SCS requires to be preserved) because [1] caused
the vDSO in the Pixel 2 and 3 kernels released with P to clobber
x18. This patch never landed in the upstream or common kernels, so
(to the best of my knowledge) only Pixel 2 and 3 are affected. A newer
version of the patch that is currently being considered for upstream
and common includes a fix for this issue [2]. The fix is also present
in the development versions of the Pixel 2 and 3 kernels.

This patch causes us to preserve x18 when we're running on a kernel
which might be one of the affected Pixel kernels (namely 4.4 or 4.9)
by replacing the vDSO function pointer with a pointer to a function
that stores x18 on the stack and then calls the vDSO. We only do this
if necessary so that we avoid leaking x18 to the stack (albeit for
a short amount of time), and because the functions in the vDSO are
performance sensitive (which after all is the whole reason why they
are in the vDSO).

[1] https://android.googlesource.com/kernel/msm/+/1bfa5df2a35bb81cd58f6c87cf3695ed880fac50
[2] https://patchwork.kernel.org/patch/10671685/

Bug: 122331452
Test: enabled SCS in libbluetooth{,-binder,_jni}.so and built aosp_walleye-userdebug, device still boots and Bluetooth works
Change-Id: I5c30fa897b1efe3a14c690004e0652adf6a2d98d