Get dns failure information
Want to obtain the abnormality of the domain name and its
abnormal information (name, uid, qtype, T_A or T_AAAA and retry),
and analyze the acquired information to discover the possible
network problems of the devices.
After the upper-layer app get the dns failure information,
it can automatically recover from the abnormal scenario.

Bug: 120624459
Change-Id: I2bab5452da91a5c1895ee60d7bacbd494a2bc458