Update patch set 4
Patch Set 4:

> Patch Set 1:
> 
> (3 comments)

[res_cache.c]
You are totally right.

I want to know if it sent DNS query to the network or just got response from cache when DNS query is failed.

For this I used netContext and it shouldn't be const.

Please review it and let me know if there is more good way to get the info.

[res_send.c]
I modified as you suggested.

Patch-set: 4
Reviewer: Gerrit User 1424063 <1424063@85c56323-6fa9-3386-8a01-6480fb634889>