Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Yeah, this basically looks good to me.
> 
> Do phdr_table_serialize_gnu_relro and phdr_table_map_gnu_relro currently work when multiple solibs are loaded? From inspection, it looks like find_libraries would call phdr_table_serialize_gnu_relro for each solib. The serialize func would write the solib's RELRO sections to the fd, but then map the start of the fd to the solib's RELRO.

You're right - whoops :)

> i.e. Loading multiple solibs with phdr_table_serialize_gnu_relro with ANDROID_DLEXT_WRITE_RELRO would corrupt the RELRO memory for the second and later solibs. Maybe I'm missing something though.

It won't actually corrupt it, because phdr_table_map_gnu_relro is extremely careful to only map over pages that are actually identical. So, the second and later solibs will be fine, but will not actually share any RELRO memory (defeating the point).

> It seems easy to fix -- find_libraries could pass the address of a "relro_fd_offset" to each call to link_image.

Yup, that should work.

> (Sorry for the delay in reviewing the change.)

No worries, I've been busy with many other things so haven't had time to come back to it either.

I'm trying out that fix and trying to write a test that covers the RELRO sharing case. It's tricky, though: it's easy to test that it doesn't crash/corrupt anything, but it's harder to test that the memory sharing actually *happens* - the current test for this just relies on doing a side by side comparison and counting the number of pages, and assuming that sharing will save at least one page. That approach doesn't extend to multiple .so files because there was at least one page of sharing even when it wasn't working properly (from the first library).

Patch-set: 1
Reviewer: Gerrit User 1021004 <1021004@85c56323-6fa9-3386-8a01-6480fb634889>