Add prebuilts that serves as mount points for bionic libs and linker
Bootstrap bionic (the bionic in the system partition that is for the
early processes before the runtime APEX is ready) is now located at
* /system/lib[64]/bootstrap/lib{c|dl|m}.so
* /system/bin/bootstrap/linker[64]

At runtime, by the init process, they are bind-mounted to the usual
paths of /system/lib[64]/lib{c|dl|m}.so and /system/bin/linker[64].
In order to do the bind mounts, there must be files that will serve as
mount points. This change adds that files as prebuilts. The files are
automatically installed when the bootstrap bionic is installed.

Bug: 120266448
Test: m libc.bootstrap libm.bootstrap libdl.bootstrap linker.bootstrap
Following files are found under /system
./lib[64]/libc.so, libdl.so, libm.so (all dummy files)
./lib[64]/bootstrap/libc.so, libdl.so, libm.so (all real so files)
./bin/linker, linker64 (dummy files)
./bin/linker_asan, linker_asan64 (symlinks)
./bin/bootstrap/linker, linker64 (real elf files)
./bin/bootstrap/linker_asan, linker_asan64 (symlinks)

Change-Id: Iab9ce94530a893950bb0b440b664e25227681889