Update patch set 15
Patch Set 15:

> Patch Set 15:
> 
> This CL in conjunction with I09b78e38df285033ef6e9c85f7ea4b0274e85070 breaks my WIP prebuilt ELF checker in a subtle way.  Because I09b78e38df285033ef6e9c85f7ea4b0274e85070 simply removes libc, libdl, and libm from LOCAL_SHARED_LIBRARIES thus the checker always emits errors on stale DT_NEEDED entries (a.k.a. prebuilt binaries are using libc.so, libm.so, and libdl.so that no longer exist).

Oh, that's unfortunate but I think we are doing it right. The change removes a lib from LOCAL_SHARED_LIBRARIES if the lib is a stubs library which is in an APEX. Otherwise, we will have duplicated copy of the lib (one in the system partition, the other in the APEX) even though the system copy isn't used by anyone.

Patch-set: 15