Add bionic mount points under /bionic
This change adds following files and symlinks:

Files:
/bionic/lib[64]/lib{c|dl|m}.so
/bionic/bin/linker[64]

Symlinks:
/system/lib[64]/lib{c|dl|m}.so -> /bionic/lib[64]/lib{c|dl|m}.so
/system/bin/linker[64] -> /bionic/bin/linker[64]

The files serve as mount points for either bootstrap bionic or the bionic
from the runtime APEX by the init. The symlinks are there to not change
the ordinary paths to the bionic files; there are many places that the paths
are implied or hard-coded, e.g., dlopen("/system/lib/libc.so") or
DT_INTERP pointing to /system/bin/linker in the vendor prebuilts.

Bug: 120266448
Test: m
The aforementioned files and symlinks are found

Change-Id: I97e38c29409ac0610dde285db8df6e94a7930094