sigpending: fix crash upon invalid sigset_t pointer
When passing an invalid pointer for sigset_t* to `sigpending`, the
syscall is diagnosing this error with returning -1 and setting errno
EFAULT.

The current implementation of the syscall wrapper is assigning the value
of `bionic_set` withough prior validation to a local variable, causing a
SIGSEGV upon dereferencing an invalid `bionic_set`.

In order to restore the syscall behaviour, cast the sigset_t* to a
sigset64_t* to achieve the same effect as using the SigSetConverter.

Test: LTP/sigpending02
Test: bionic-unit-tests
Bug: 112484513
Change-Id: I0633db1ff526584b244184abf9f5eb3bf1511269