Greylist libraries are attempted to be loaded in the default ns
Some of the greylist libraries can be in directories other than
/system/lib. For example, libnativehelper.so is moved from the directory
to /apex/com.android.runtime/lib.

Previous behavior for greylist libraries is to find the lib under
/system/lib while "within" the current namespace which is usually the
classloader-namespace. Since the libs no longer exist there, linker
fails to load it and no further attempt is made.

This change fixes the problem by giving a second chance to load the
greylist lib "from" the default namespace. Since there is a link from the
default namespace to the namespace where the moved libraries are
intended to be loaded (e.g. the 'runtime' namespace for
libnativehelper.so) the call is successful.

Bug: 124201397
Test: WIP

Change-Id: I11d15993d18896bdf663c408f9e40d8a607c9b80