Export android_getaddrinfofornetcontext for netd
Back in P, netd called this symbol from DnsProxyListener.cpp. Now netd
could direcly call into libnetd_resolv.so, but for the sake of reducing
the ABI surface, it's still preferable to take the longer toop via
bionic's dns proxy. This is what we already do for all the other
resolver entry point, and this one should be no exception.

Test: atest netd_integration_test (with patch to hide the symbol in libnetd_resolv)
Change-Id: Ie68826c76354741bb41009298daf385661456b65