Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> does +cferris need to do anything in Android's unwinder?

I tested that tombstone unwinding works correctly for PT_GNU_EH_FRAME. I haven't tested it with PT_ARM_EXIDX but given that libunwindstack is already doing the right thing in common code [1,2] I would also expect it to work. That was the next thing that I was going to try once I got lld to create a PT_ARM_EXIDX for the other partition as well, but now that I think about it I could just use something like in the test that I'm adding here to test it. It seems worth adding a test to libunwindstack in any case, though.

[1] https://android.googlesource.com/platform/system/core/+/master/libunwindstack/ElfInterface.cpp#51
[2] https://android.googlesource.com/platform/system/core/+/master/libunwindstack/ElfInterface.cpp#213

Patch-set: 1
Reviewer: Gerrit User 1067098 <1067098@85c56323-6fa9-3386-8a01-6480fb634889>