Update patch set 1
Patch Set 1:

I will abandon this CL.

In fact, we don't need this at all.  The secret is to implement a small set of system calls as (local) static functions and use those functions to read device information.  The code in [1] is a good reference.

[1] https://android.googlesource.com/platform/bionic/+/e4788d4c7e842b0ca24363ebc5544a16c9e3b185/libc/arch-arm/dynamic_function_dispatch.cpp#66

Patch-set: 1
Reviewer: Gerrit User 1014807 <1014807@85c56323-6fa9-3386-8a01-6480fb634889>