Update patch set 3
Patch Set 3: Code-Review+2

I am fine with the order in this change.

At least for APEXes, we try to have an executable only in one place; for example, dex2oat was moved to the runtime APEX. We will make sure that the binary is not installed to any other places like system or product.

(We still have a symlink to it in the system partition. However, that's not to override it, but to provide backward compatibility.)

Patch-set: 3
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2