clean_header: Fix error handling for no-such-file case
If the function cleanupFile encounters an error, it is expect to return
a value that evaluates to False. As it stands, however, it returns
(None, None) which does not evaluate to False. Change this function to
return False, if it encounters an error.

We previously saw the following error message, when we tried to run
clean_header.py on a non-existant file.

Traceback (most recent call last):
  File "clean_header.py", line 208, in <module>
    b.updateGitFiles()
  File "utils.py", line 164, in updateGitFiles
    self._writeFile(dst)
  File "utils.py", line 136, in _writeFile
    f.write(self.new_data[dst])
TypeError: expected a string or other character buffer object

Change-Id: I5f717dd1a4388f598f0fd4bfd5e6129017de9095