Update patch set 1
Patch Set 1:

Is this a bug with strace, not with the seccomp filter? For an unprivileged app to succeed at calling setreuid(), it must be the case that the UID we're attempting to change to is the UID we're already running as. If so, then there's no need to make the setreuid() call.

It feels like the bug here isn't setcomp, but rather, strace, and we should just fix strace to not call setreuid.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>