Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Is this a bug with strace, not with the seccomp filter? For an unprivileged app to succeed at calling setreuid(), it must be the case that the UID we're attempting to change to is the UID we're already running as. If so, then there's no need to make the setreuid() call.
> 
> It feels like the bug here isn't setcomp, but rather, strace, and we should just fix strace to not call setreuid.

That doesn't seem like a bug in strace, though? If strace is running as root, it makes sense to ptrace the process and then reduce privileges to match the traced process. Why should they check whether the uid matches when the kernel does it for you when you call setreuid?

Patch-set: 1
CC: Gerrit User 1079148 <1079148@85c56323-6fa9-3386-8a01-6480fb634889>