Update patch set 5
Patch Set 5: Presubmit-Ready+1

ps5 tests for the enumeration constants in the header test, reorders the functions to match the order in the C11 standard (i had another look through the 2018 edition to triple-check, and that was all i noticed).

it also fixes the thrd_exit test --- it's explicitly defined that EXIT_SUCCESS is returned if you're the last thread. that's the behavior we were getting from pthread_exit, which is why this failed presubmit. i'd accidentally not run this test previously because my gtest filter of '*thread.*' wasn't matching thread_DeathTest...

i've added a test and improved the comments based on a reading of the 2018 version of the standard, which includes the fix for the defect reports about tss_t destructors being underspecified.

Patch-set: 5
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1