tests: use clang's -verify instead of FileCheck
Clang uses this for its own diagnostic tests. The nice feature it adds
here is that it fails if clang emits a diagnostic that we don't expect
(and we get to drop a dependency on FileCheck).

This also includes a migration from Android.mk to Android.bp. It's a
little bit fragile, since `-Xclang -verify` means that clang *doesn't*
produce any artifacts. Still, it seems that the build system is happy to
cope with that (successive `mma`s don't rerun this test until
fortify_filecheck_diagnostics_test.cpp is updated)

Bug: 131861088
Test: mma
Change-Id: I77f7ce77869edaa23e2401e622ad7007d2fee06c